Some Package restructuring
