+++ pom.xml
@@ -5,7 +5,7 @@
 	<modelVersion>4.0.0</modelVersion>
 
 	<groupId>snake_game</groupId>
-	<artifactId>aiaiaia</artifactId>
+	<artifactId>ASD_SnakeGame</artifactId>
 
 	<version>1.0-SNAPSHOT</version>
 	<packaging>jar</packaging>
@@ -21,7 +21,7 @@
 					</descriptorRefs>
 					<archive>
 						<manifest>
-							<mainClass>sample.src.sample.GameLoop</mainClass>
+							<mainClass>GameLoop</mainClass>
 						</manifest>
 					</archive>
 				</configuration>
@@ -38,6 +38,7 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
+				 <version>3.8.1</version>
 				<configuration>
 					<source>6</source>
 					<target>6</target>
