+++ src/main/java/asd/Control.java
@@ -1,5 +1,6 @@
-package sample.src.sample;
+package asd;
 
+import asd.GameLoop;
 import javafx.scene.Group;
 import javafx.scene.input.KeyEvent;
 import javafx.stage.Stage;
@@ -8,7 +9,7 @@ public class Control {
 
     private boolean goUp, goDown, goRight, goLeft;
 
-    void stopMovement() {
+    public void stopMovement() {
         goUp = false;
         goDown = false;
         goRight = false;
