+++ src/main/java/asd/EatenTailMessage.java
@@ -1,4 +1,4 @@
-package sample.src.sample;
+package asd;
 
 import java.util.ArrayList;
 import java.util.Collection;
