+++ src/main/java/asd/GameObject.java
@@ -1,4 +1,4 @@
-package sample.src.sample;
+package asd;
 
 import javafx.geometry.Bounds;
 import javafx.scene.Group;
@@ -14,11 +14,11 @@ public class GameObject {
     private double redPart, greenPart, bluePart;
     private Bounds fbound;
 
-    public GameObject() {
+//    public GameObject() {
+//
+//    }
 
-    }
-
-    double[] getColor() { // returned ein double Array mit den Farben fÃ¼r den Schwanz der Schlange, wird nacher von eat aufgerufen
+    double[] getColor() { // returned ein double Array mit den Farben für den Schwanz der Schlange, wird nacher von eat aufgerufen
         double[] colors = new double[3];
         colors[0] = redPart;
         colors[1] = greenPart;
@@ -32,7 +32,7 @@ public class GameObject {
         groupObject.getChildren().remove(food);//um vorheriges Food verschwinden zu lassen
         Random rand = new Random();
 
-        food.setFill(Color.color(redPart = rand.nextDouble(), greenPart = rand.nextDouble(), bluePart = rand.nextDouble())); // hier werden zufÃ¤llige Farben fÃ¼r das Food (und damit auch den Tail) Ã¼bergeben
+        food.setFill(Color.color(redPart = rand.nextDouble(), greenPart = rand.nextDouble(), bluePart = rand.nextDouble())); // hier werden zufällige Farben für das Food (und damit auch den Tail) übergeben
         food.relocate(rand.nextInt((int) stageObject.getWidth() - 50), rand.nextInt((int) stageObject.getHeight() - 50)); // Random Location mit Abstand vom Rand jeweils 40
         groupObject.getChildren().add(food);
         fbound = food.getBoundsInParent();
